Moved credentials package down to studio/
